refactor: 성별, 활동량, 당 섭취 목표 enum을 CommonModel로 이동하여 공통 사용
