+++ HumanNeverDie/Projects/Features/SettingFeature/Sources/SettingModifier.swift
@@ -1,0 +1,18 @@
+//
+//  SettingModifier.swift
+//  SettingFeature
+//
+//  Created by Seulki Lee on 7/17/25.
+//
+
+import SwiftUI
+
+struct SettingModifier: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+#Preview {
+    SettingModifier()
+}
