+++ HumanNeverDie/Projects/Features/SettingFeature/Sources/StepsView/AccountInfoView.swift
@@ -1,0 +1,18 @@
+//
+//  accountInfoView.swift
+//  SettingFeature
+//
+//  Created by Seulki Lee on 7/16/25.
+//
+
+import SwiftUI
+
+struct accountInfoView: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+#Preview {
+    accountInfoView()
+}
