+++ HumanNeverDie/Projects/Features/SettingFeature/Sources/StepsView/AccountInfoViewFactory.swift
@@ -1,0 +1,18 @@
+//
+//  SwiftUIView.swift
+//  SettingFeature
+//
+//  Created by Seulki Lee on 7/17/25.
+//
+
+import SwiftUI
+
+struct SwiftUIView: View {
+    var body: some View {
+        Text(/*@START_MENU_TOKEN@*/"Hello, World!"/*@END_MENU_TOKEN@*/)
+    }
+}
+
+#Preview {
+    SwiftUIView()
+}
