+++ HumanNeverDie/Projects/Features/SettingFeature/Sources/StepsView/AccountInfoViewModel.swift
@@ -1,0 +1,8 @@
+//
+//  AccountInfoViewModel.swift
+//  SettingFeature
+//
+//  Created by Seulki Lee on 7/16/25.
+//
+
+import Foundation
